Backend/bridge-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
aiosqlite
paho-mqtt
python-multipart
pycryptodome
